requirements.txt
python-Levenshtein
matplotlib==3.5.0
numpy
pandas
tqdm
openpyxl
pdf2image